Move DashboardTitleContainer to components folder
