+++ packages/lesmis/src/components/DashboardExportModal/components/index.js
@@ -1,3 +1,4 @@
 export { EntityDetails } from './EntityDetails';
 export { A4Page } from './A4Page';
 export { OptionsBar } from './OptionsBar';
+export { DashboardTitleContainer } from './DashboardTitleContainer';
